+++ lib/karafka/web/pro/ui/lib/search/matchers/raw_payload_includes.rb
@@ -19,9 +19,13 @@ module Karafka
           module Search
             module Matchers
               # Checks for phrase existence in the raw payload
+              # If raw payload has a `zlib` header, it tries to decompress it
+              # (without de-serializing). We decompress because internal web ui topics (including
+              # errors topic) use Ruby specific zlib compression.
               #
               # @note It is case sensitive
               # @note Ignores encoding issues
+              # @note Decompresses if zlib is used
               class RawPayloadIncludes < Base
                 # @param message [Karafka::Messages::Message]
                 # @param phrase [String]
@@ -30,10 +34,26 @@ module Karafka
                   # raw payload can be nil for tombstone events
                   return false unless message.raw_payload
 
-                  message.raw_payload.include?(phrase)
+                  build_matchable_payload(message).include?(phrase)
                 # String matching on compressed data may fail
                 rescue Encoding::CompatibilityError
                   false
+                end
+
+                private
+
+                # Checks whether decompression is needed and if yes does it. Does not deserialize
+                #   the payload itself.
+                # @param message [Karafka::Messages::Message]
+                # @return [String] decompressed payload
+                def build_matchable_payload(message)
+                  raw_payload = message.raw_payload
+
+                  return raw_payload unless message.raw_headers.key?('zlib')
+
+                  Zlib::Inflate.inflate(raw_payload)
+                rescue Zlib::Error
+                  raw_payload
                 end
               end
             end